URL path alignment and CSS class change
In the `ChangeCulture.razor` file, changed the CSS class `MudDivider` from `mt-1 mb-2` to `my-2`, which probably changes the margins of the element.

In the `ChangeOrganization.razor` file, the navigation path in the `GoToManage` method was changed from `“Organizations”` to `“organizations”`, which may be related to URL path alignment.

In the `ChangeProjects.razor` file, the navigation path in the `GoToManage` method was changed from `“Projects”` to `“projects”`, which may also be related to URL path alignment.